feat(404 pages): Refactor not found page for general, landelijk, and articles.
